Make resource-processor a web service now - change structure a bit - add a new field to dto AudioFileMetadata: resourceId and adjust mapper - refactor MetadataService - add application props - minor fixes
