+++ resource-processor/src/main/java/com/epam/microservices/resourceprocessor/exception/ExtractMetadataException.java
@@ -1,0 +1,8 @@
+package com.epam.microservices.resourceprocessor.exception;
+
+public class ExtractMetadataException extends RuntimeException {
+
+    public ExtractMetadataException(Throwable cause) {
+        super(cause);
+    }
+}
