+++ resource-processor/src/main/java/com/epam/microservices/resourceprocessor/model/AudioFileMetadata.java
@@ -1,18 +1,21 @@
-package com.epam.microservices.model;
+package com.epam.microservices.resourceprocessor.model;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import lombok.With;
 
 @Getter
 @Setter
 @Builder
 @ToString
 public class AudioFileMetadata {
-    private String songName;
+    private String name;
     private String artist;
     private String album;
     private String length;
     private Integer year;
+    @With
+    private Integer resourceId;
 }
